% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions.back202502.R
\name{get_ect_systemfit}
\alias{get_ect_systemfit}
\title{Extract Error Correction Term (ECT)}
\usage{
get_ect_systemfit(
  systemfit_uecm_coefs,
  sel_variables,
  table_dt,
  nperiods,
  nunits
)

get_ect_systemfit(
  systemfit_uecm_coefs,
  sel_variables,
  table_dt,
  nperiods,
  nunits
)
}
\arguments{
\item{systemfit_uecm_coefs}{A list containing coefficients from UECM.}

\item{sel_variables}{A character vector of selected variable names.}

\item{table_dt}{Dataframe of origin for all variables.}

\item{nperiods}{An integer specifying the number of time observations by unit.}

\item{nunits}{An integer reflecting the number of panel units}
}
\value{
A data.table with the ECT.

A data.table object containing the error correction term.
}
\description{
Computes the ECT from UECM coefficients.

Computes the Error Correction Term from Unrestricted ECM coefficients.
}
