% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions.back202502.R
\name{systemfit_boundsF_test}
\alias{systemfit_boundsF_test}
\title{Perform Bounds F-Test}
\usage{
systemfit_boundsF_test(system_ecm, units, sel_variables)

systemfit_boundsF_test(system_ecm, units, sel_variables)
}
\arguments{
\item{system_ecm}{An object of class systemfit, representing the ECM.}

\item{units}{A character vector specifying the units or entities for the model.}

\item{sel_variables}{Variables included in the cointegration relationship}
}
\value{
Numeric vector of F-test results.

A numeric vector with F-test results for each unit.
}
\description{
Performs the Bounds F-test.

Applies the Bounds F-Test to the system equations based on Pesaran (2001).
}
