% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{uecm_systemfit}
\alias{uecm_systemfit}
\title{Estimate an Unrestricted Error Correction Model (UECM) using systemfit}
\usage{
uecm_systemfit(
  data,
  col_names,
  nlags = 1,
  grouping,
  method = "SUR",
  method_solv = "EViews",
  iterations = 1,
  inst_list = NULL
)
}
\arguments{
\item{data}{A data.frame or data.table.}

\item{col_names}{Character vector of variable names.}

\item{nlags}{Integer, number of lags.}

\item{grouping}{Character string, name of the grouping column.}

\item{method}{"SUR", "2SLS", or "3SLS".}

\item{method_solv}{Solution method for 3SLS (see \code{systemfit.control}).}

\item{iterations}{Integer, max iterations.}

\item{inst_list}{Character vector of instruments (for 2SLS/3SLS).}
}
\value{
A \code{systemfit} object.
}
\description{
Estimates a UECM for panel data using \code{systemfit}.
}
