% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions2.R
\name{get_ect_systemfit}
\alias{get_ect_systemfit}
\title{Extract Error Correction Term from Systemfit Model}
\usage{
get_ect_systemfit(systemfit_uecm, sel_variables, data, grouping)

get_ect_systemfit(systemfit_uecm, sel_variables, data, grouping)
}
\arguments{
\item{systemfit_uecm}{A fitted model object of class `systemfit` from
`uecm_systemfit`.}

\item{sel_variables}{A character vector of selected variable names *in
levels* (not differenced) that form the cointegrating relationship. The
first variable *must* be the dependent variable in levels.}

\item{data}{A `data.table` object containing the original data (must include
the `grouping` columns).}

\item{grouping}{A character vector of length two, specifying the columns
defining panel units and time periods (e.g., `c("reporter", "year")`).}

\item{systemfit_uecm_coefs}{A list containing coefficients from UECM.}

\item{nperiods}{An integer specifying the number of time observations by unit.}

\item{nunits}{An integer reflecting the number of panel units}

\item{table_dt}{Dataframe of origin for all variables.}
}
\value{
A data.table object containing the error correction term.

A `data.table` object containing the error correction term.
}
\description{
Computes the Error Correction Term from Unrestricted ECM coefficients.

Computes the Error Correction Term (ECT) from coefficients of an
Unrestricted Error Correction Model (UECM) estimated using `systemfit`.
}
