% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_ect_systemfit}
\alias{get_ect_systemfit}
\title{Extract Error Correction Term (ECT)}
\usage{
get_ect_systemfit(
  systemfit_uecm_coefs,
  sel_variables,
  table_dt,
  nperiods,
  nunits
)
}
\arguments{
\item{systemfit_uecm_coefs}{\code{systemfit} object (UECM results).}

\item{sel_variables}{Character vector of variable names.}

\item{table_dt}{Data.table containing the original data.}

\item{nperiods}{Number of time periods.}

\item{nunits}{Number of panel units.}
}
\value{
A data.table with the ECT.
}
\description{
Computes the ECT from UECM coefficients.
}
