% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions2.R
\name{recm_systemfit}
\alias{recm_systemfit}
\title{Restricted Error Correction Model for Systemfit}
\usage{
recm_systemfit(
  formula,
  data,
  grouping,
  uecm_model,
  nlags = 1,
  method = "SUR",
  method_solv = "EViews",
  iterations = 1,
  inst_list = NULL
)

recm_systemfit(
  formula,
  data,
  grouping,
  uecm_model,
  nlags = 1,
  method = "SUR",
  method_solv = "EViews",
  iterations = 1,
  inst_list = NULL
)
}
\arguments{
\item{formula}{A formula of the form `y ~ x1 + x2 | z1 + z2`, where `y` is the
dependent variable, `x1` and `x2` are independent variables, and `z1` and
`z2` are instrumental variables (optional).  The formula *must* be
explicitly specified.}

\item{data}{A `data.table` containing the original data.}

\item{grouping}{A character vector of length two specifying the panel unit
and time period columns (e.g., `c("reporter", "year")`).}

\item{uecm_model}{A fitted model object of class `systemfit` representing the
precomputed UECM model. This is used to calculate the ECT.}

\item{nlags}{An integer specifying the number of lags included in the UECM.
Defaults to 1. This should match the `nlags` used in `uecm_systemfit`.}

\item{method}{A character string indicating the desired estimation method.
Must be one of "SUR" (default), "OLS", "2SLS", or "3SLS".}

\item{method_solv}{A character string indicating the solution method for 3SLS.
Default is "EViews".}

\item{iterations}{An integer indicating the number of iterations. Defaults to 1.}

\item{inst_list}{A character vector of instruments for 2SLS and 3SLS methods.  See
`uecm_systemfit` for details.}

\item{col_names}{A character vector of column names.}

\item{nunits}{An integer reflecting the number of panel units}

\item{nperiods}{An integer specifying the number of time observations by unit.}

\item{dt}{A data.table object containing the data.}
}
\value{
A model result from systemfit.

A model result from `systemfit`.
}
\description{
Computes the Restricted Error Correction Model for Systemfit.

Computes the Restricted Error Correction Model (RECM) using a
precomputed UECM model and an extracted Error Correction Term (ECT).
}
