% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{recm_systemfit}
\alias{recm_systemfit}
\title{Restricted Error Correction Model for Systemfit}
\usage{
recm_systemfit(
  col_names = c(),
  uecm_model,
  method = "SUR",
  method_solv = "EViews",
  iterations = 1,
  nlags = 1,
  dt = data.table::data.table(),
  inst_list = c()
)
}
\arguments{
\item{col_names}{A character vector of column names.}

\item{uecm_model}{An object of class systemfit, representing the UECM model.}

\item{method}{Character string indicating the desired estimation method.}

\item{method_solv}{Character string indicating the solution method. Default is "EViews".}

\item{iterations}{An integer indicating the number of iterations.}

\item{nlags}{An integer specifying the number of lags.}

\item{dt}{A data.table object containing the data.}

\item{inst_list}{List of instruments for 2SLS and 3SLS.}
}
\value{
A model result from systemfit.
}
\description{
Computes the Restricted Error Correction Model for Systemfit.
}
