% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions2.R
\name{systemfit_boundsF_test}
\alias{systemfit_boundsF_test}
\title{Bounds F-Test for Systemfit Error Correction Model}
\usage{
systemfit_boundsF_test(system_ecm, sel_variables)

systemfit_boundsF_test(system_ecm, sel_variables)
}
\arguments{
\item{system_ecm}{An object of class `systemfit`, representing the Error
Correction Model (ECM), either UECM or RECM.  The test is applied to the
first equation in the system.}

\item{sel_variables}{A character vector of variables *in levels* included in
the cointegration relationship. The first variable must correspond to the
dependent variable in the ECM.}

\item{units}{A character vector specifying the units or entities for the model.}
}
\value{
A numeric vector with F-test results for each unit.

A single numeric value representing the F-statistic.
}
\description{
Applies the Bounds F-Test to the system equations based on Pesaran (2001).

Applies the Bounds F-Test (Pesaran, Shin, and Smith, 2001) to a single equation of the
system equations estimated by `systemfit`. This tests for cointegration.
}
