% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{systemfit_boundsF_test}
\alias{systemfit_boundsF_test}
\title{Perform Bounds F-Test}
\usage{
systemfit_boundsF_test(system_ecm, units, sel_variables)
}
\arguments{
\item{system_ecm}{\code{systemfit} object (ECM results).}

\item{units}{Character vector of unit names.}

\item{sel_variables}{Character vector of variables in cointegration relationship.}
}
\value{
Numeric vector of F-test results.
}
\description{
Performs the Bounds F-test.
}
