% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/functions.back202502.R
\name{uecm_systemfit}
\alias{uecm_systemfit}
\title{Estimate an Unrestricted Error Correction Model (UECM) using systemfit}
\usage{
uecm_systemfit(
  col_names = c(),
  nlags = 1,
  grouping,
  method = "SUR",
  method_solv = "EViews",
  iterations = 1,
  dt = data.table::data.table(),
  inst_list = c()
)

uecm_systemfit(
  col_names = c(),
  nlags = 1,
  grouping,
  method = "SUR",
  method_solv = "EViews",
  iterations = 1,
  dt = data.table::data.table(),
  inst_list = c()
)
}
\arguments{
\item{col_names}{A character vector of column names.}

\item{nlags}{An integer specifying the number of lags.}

\item{grouping}{Column defining panel units.}

\item{method}{Character string indicating the desired estimation method.}

\item{method_solv}{Character string indicating the solution method. Default is "EViews".}

\item{iterations}{An integer indicating the number of iterations.}

\item{dt}{A data.table object containing the data.}

\item{inst_list}{List of instruments for 2SLS and 3SLS.}

\item{data}{A data.frame or data.table.}
}
\value{
A \code{systemfit} object.

A model result from systemfit.
}
\description{
Estimates a UECM for panel data using \code{systemfit}.
}
\examples{
# Example (replace with your data)
# data(your_data)
# uecm_model <- uecm_systemfit(data = your_data,...)
}
